backend/requirements.txt
fastapi
uvicorn
groq
openai
python-dotenv
pydantic
